Refactor database folder resolution
As we've transitioned the to-do system into a proper application, the previous
method of modifying the database folder location in `test_helper.exs` is no
longer viable. This happens because `test_helper.exs` runs after the application
has already started, meaning the internal database folder has already been
resolved.

A better approach is to use `config/runtime.exs`, which is evaluated during
release startup. This allows us to use the system environment as it exists on
the target system.