Optimize cache `server_process/1` for todo-server global registration
Global process registration is much more expensive than a local registration, as
it requires synchronized communication across the entire cluster. In the current
implementation, even if the server process is already running, we attempt a
`:global` registration every time we want to work with a to-do list.

This commit addresses this issue by performing an explicit lookup first.